fix: 프레인 선택 스크롤 - 컴포넌트 분리
