+++ PawCut/PawCut/Sources/Views/PawcutCamera/PawcutFrameView.swift
@@ -32,46 +32,17 @@ struct PawcutFrameView: View {
                 }
 
                 Spacer()
-                
+
                 PawTitleLabel.bold18(viewModel.selectedFrameDisplayName ?? "")
                     .padding(.horizontal, 21)
 
-                ScrollView(.horizontal, showsIndicators: false) {
-                    HStack(spacing: 8) {
-                        ForEach(
-                            // TODO: 왜 Array 써야하지?
-                            Array(viewModel.frameScrollImageNames.enumerated()),
-                            id: \.offset
-                        ) { index, name in
-
-                            let image = UIImage(named: name)
-                            let isSelected =
-                                index == viewModel.selectedFrameIndex
-
-                            if let image = image {
-                                Image(uiImage: image)
-                                    .resizable()
-                                    .frame(width: 40, height: 40)
-                                    .clipShape(Circle())
-                                    .overlay(
-                                        Circle()
-                                            .strokeBorder(
-                                                isSelected
-                                                    ? Color.pointPurple01
-                                                    : .clear,
-                                                lineWidth: 2
-                                            )
-                                    )
-                                    .onTapGesture {
-                                        viewModel.selectedFrameIndex = index
-                                    }
-                            }
-                        }
+                PawcutFrameSelectionScrollView(
+                    frameNames: viewModel.frameScrollImageNames,
+                    selectedFrameIndex: viewModel.selectedFrameIndex,
+                    onSelect: { index in
+                        viewModel.selectedFrameIndex = index
                     }
-                    .padding(.horizontal)
-                }
-                .padding(.top, 31)
-                .padding(.leading, 4)
+                )
 
                 PawPrimaryButton("저장하기") {
                     let exportView = PawcutGridImagePreview(
