+++ PawCut/PawCut/Sources/Views/PawcutCamera/SubViews/PawcutFrameSelectionScrollView.swift
@@ -1,0 +1,54 @@
+//
+//  PawcutFrameSelectionScrollView.swift
+//  PawCut
+//
+//  Created by Jay on 9/23/25.
+//
+
+import SwiftUI
+
+
+struct PawcutFrameSelectionScrollView: View {
+    let frameNames: [String]
+    let selectedFrameIndex: Int?
+    let onSelect: (Int) -> Void
+    
+    var body: some View {
+        ScrollView(.horizontal, showsIndicators: false) {
+            HStack(spacing: 8) {
+                ForEach(
+                    // TODO: 왜 Array 써야하지?
+                    Array(frameNames.enumerated()),
+                    id: \.offset
+                ) { index, name in
+
+                    let image = UIImage(named: name)
+                    let isSelected =
+                        index == selectedFrameIndex
+
+                    if let image = image {
+                        Image(uiImage: image)
+                            .resizable()
+                            .frame(width: 40, height: 40)
+                            .clipShape(Circle())
+                            .overlay(
+                                Circle()
+                                    .strokeBorder(
+                                        isSelected
+                                            ? Color.pointPurple01
+                                            : .clear,
+                                        lineWidth: 2
+                                    )
+                            )
+                            .onTapGesture {
+                                onSelect(index)
+                            }
+                    }
+                }
+            }
+            .padding(.horizontal)
+        }
+        .padding(.top, 31)
+        .padding(.leading, 4)
+    }
+}
